OpenGL Integration, Termination of Double, bug fixes
OpenGL will be the second front to game rendering system along side Java2D. Every instance of double has turned into float, to shrink excess precision from game's memory usage and  easier delegating to Renderers. Shortened the TH20 GST.

+++ src/action/CompareValue.java
@@ -93,8 +93,8 @@ public final class CompareValue {
 
 		if (allNumbers) {
 			for (int i = 0; i < resolved.length - 1; i++) {
-				double a = ((Number) resolved[i]).doubleValue();
-				double b = ((Number) resolved[i + 1]).doubleValue();
+				float a = ((Number) resolved[i]).floatValue();
+				float b = ((Number) resolved[i + 1]).floatValue();
 
 				if (!compareNumbers(a, b, comparison)) {
 					return false;
@@ -111,8 +111,8 @@ public final class CompareValue {
 	}
 
 	private static boolean compareNumbers(
-		double a,
-		double b,
+		float a,
+		float b,
 		Comparison comparison
 	) {
 		return switch (comparison) {
